+++ nodes/uibuilder.js
@@ -331,17 +331,19 @@ module.exports = function(RED) {
             node.ioClientsCount++
 
             log.debug(
-                `UIbuilder: ${node.url} Socket connected, clientCount: ${node.ioClientsCount}, ID: ${socket.id}, Cookie: ${socket.handshake.headers.cookie}`
+                `UIbuilder: ${node.url} Socket connected, clientCount: ${node.ioClientsCount}, ID: ${socket.id}`
             )
+
             setNodeStatus( { fill: 'green', shape: 'dot', text: 'connected ' + node.ioClientsCount }, node )
 
-            // Let the clients know we are connecting & send the desired debug state
-            ioNs.emit( node.ioChannels.control, { 'type': 'server connected', 'debug': node.debugFE } )
+            // Let the clients (and output #2) know we are connecting & send the desired debug state
+            sendControl({ 'type': 'server connected', 'debug': node.debugFE, '_socketId': socket.id }, ioNs, node)
+            //ioNs.emit( node.ioChannels.control, { 'type': 'server connected', 'debug': node.debugFE } )
 
             // if the client sends a specific msg channel...
             socket.on(node.ioChannels.client, function(msg) {
                 log.debug(
-                    `UIbuilder: ${node.url}, Data received from client, ID: ${socket.id}, Cookie: ${socket.handshake.headers.cookie}, Msg: ${msg.payload}`
+                    `UIbuilder: ${node.url}, Data received from client, ID: ${socket.id}, Msg: ${msg.payload}`
                 )
 
                 // Make sure the incoming msg is a correctly formed Node-RED msg
@@ -361,19 +363,47 @@ module.exports = function(RED) {
                 // TODO: This should probably have safety validations!
                 node.send(msg)
             });
+            socket.on(node.ioChannels.control, function(msg) {
+                log.debug(
+                    `UIbuilder: ${node.url}, Control Msg from client, ID: ${socket.id}, Msg: ${msg.payload}`
+                )
+
+                // Make sure the incoming msg is a correctly formed Node-RED msg
+                switch ( typeof msg ) {
+                    case 'string':
+                    case 'number':
+                    case 'boolean':
+                        msg = { 'type': msg }
+                }
+
+                if ( ! msg.hasOwnProperty('topic') ) msg.topic = node.topic
+
+                // If the sender hasn't added msg._clientId, add the Socket.id now
+                if ( ! msg.hasOwnProperty('_socketId') ) {
+                    msg._socketId = socket.id
+                }
+
+                // Send out the message on port #2 for downstream flows
+                node.send([null,msg])
+            });
 
             socket.on('disconnect', function(reason) {
                 node.ioClientsCount--
                 log.debug(
-                    `UIbuilder: ${node.url} Socket disconnected, clientCount: ${node.ioClientsCount}, Reason: ${reason}, ID: ${socket.id}, Cookie: ${socket.handshake.headers.cookie}`
+                    `UIbuilder: ${node.url} Socket disconnected, clientCount: ${node.ioClientsCount}, Reason: ${reason}, ID: ${socket.id}`
                 )
-                setNodeStatus( { fill: 'green', shape: 'ring', text: 'connected ' + node.ioClientsCount }, node )
+                if ( node.ioClientsCount <= 0) setNodeStatus( { fill: 'blue', shape: 'dot', text: 'connected ' + node.ioClientsCount }, node )
+                else setNodeStatus( { fill: 'green', shape: 'ring', text: 'connected ' + node.ioClientsCount }, node )
+                // Let the control output port know
+                node.send([null, {'type': 'client disconnect', '_socketId': socket.id, 'reason': reason, 'topic': node.topic}])
             })
 
             socket.on('error', function(err) {
                 log.error(
                     `UIbuilder: ${node.url} ERROR received, ID: ${socket.id}, Reason: ${err.message}`
                 )
+                // Let the control output port know
+                node.send([null, {'type': 'socket error', '_socketId': socket.id, 'error': err.message, 'topic': node.topic}])
             })
 
             /* More Socket.IO events but we really don't need to monitor them
@@ -494,7 +524,6 @@ function inputHandler(msg, node, RED, io, ioNs, log) {
         ioNs.emit(node.ioChannels.server, msg)
     }
 
-
     log.debug('uibuilder - msg sent to front-end via ws channel, ', node.ioChannels.server, ': ', msg)
 
     if (node.fwdInMessages) {
@@ -512,8 +541,8 @@ function processClose(done = null, node, RED, ioNs, io, app, log) {
 
     setNodeStatus({fill: 'red', shape: 'ring', text: 'CLOSED'}, node)
 
-    // Let the clients know we are closing down
-    ioNs.emit( node.ioChannels.control, { 'type': 'shutdown' } )
+    // Let all the clients know we are closing down
+    sendControl({ 'type': 'shutdown' }, ioNs, node)
 
     // Disconnect all Socket.IO clients
     const connectedNameSpaceSockets = Object.keys(ioNs.connected) // Get Object with Connected SocketIds as properties
@@ -612,6 +641,26 @@ function getProps(RED,myObj,props,defaultAnswer = []) {
         }
     }
     return ans || defaultAnswer
+}
+
+/** Output a control msg
+ * Sends to all connected clients & outputs a msg to port 2
+ * @param {object} msg The message to output
+ * @param {object} ioNs Socket.IO instance to use
+ * @param {object} node The node object
+ * @param {string=} socketId Optional. If included, only send to specific client id
+ */
+function sendControl(msg, ioNs, node, socketId) {
+    if (socketId) msg._socketId = socketId
+
+    // Send to specific client if required
+    if (msg._socketId) ioNs.to(msg._socketId).emit(node.ioChannels.control, msg)
+    else ioNs.emit(node.ioChannels.control, msg)
+
+    if ( ! msg.hasOwnProperty('topic') ) msg.topic = node.topic
+
+    // copy msg to output port #2
+    node.send([null, msg])
 }
 
 // EOF